k8s: move node annotation functions to annotate.go
No functional changes are intended by this commit.

Signed-off by: Ian Vernon <ian@cilium.io>
